move env parsers to a new pkg in utils
